feat(series): review fixes
