+++ src/pages/LegacySeries/utils.ts
@@ -1,4 +1,7 @@
+import { deprecatedSeriesURL } from '#src/utils/formatting';
+import { secondsToISO8601 } from '#src/utils/datetime';
 import type { Playlist, PlaylistItem } from '#types/playlist';
+import type { EpisodeMetadata } from '#types/series';
 
 /**
  * Get an array of options for a season filter
@@ -41,4 +44,46 @@ export const getEpisodesInSeason = (episode: PlaylistItem | undefined, seriesPla
   if (!seriesPlaylist) return;
 
   return seriesPlaylist.playlist.filter((i) => i.seasonNumber === episode?.seasonNumber)?.length;
+};
+
+export const generateLegacySeriesMetadata = (seriesPlaylist: Playlist, seriesId: string | undefined) => {
+  // Use playlist for old flow and media id for a new flow
+  const seriesCanonical = `${window.location.origin}/s/${seriesId}`;
+
+  return {
+    '@type': 'TVSeries',
+    '@id': seriesCanonical,
+    name: seriesPlaylist.title,
+    numberOfEpisodes: String(seriesPlaylist.playlist.length),
+    numberOfSeasons: String(
+      seriesPlaylist.playlist.reduce(function (list, playlistItem) {
+        return !playlistItem.seasonNumber || list.includes(playlistItem.seasonNumber) ? list : list.concat(playlistItem.seasonNumber);
+      }, [] as string[]).length,
+    ),
+  };
+};
+
+export const generateLegacyEpisodeJSONLD = (
+  seriesPlaylist: Playlist,
+  episode: PlaylistItem | undefined,
+  episodeMetadata: EpisodeMetadata | undefined,
+  seriesId: string,
+) => {
+  const episodeCanonical = `${window.location.origin}${deprecatedSeriesURL({ episodeId: episode?.mediaid, seriesId })}`;
+  const seriesMetadata = generateLegacySeriesMetadata(seriesPlaylist, seriesId);
+
+  if (!episode) {
+    return JSON.stringify(seriesMetadata);
+  }
+
+  return JSON.stringify({
+    '@context': 'http://schema.org/',
+    '@type': 'TVEpisode',
+    '@id': episodeCanonical,
+    episodeNumber: episodeMetadata?.episodeNumber,
+    seasonNumber: episodeMetadata?.seasonNumber,
+    name: episode.title,
+    uploadDate: secondsToISO8601(episode.pubdate),
+    partOfSeries: seriesMetadata,
+  });
 };
